local_requirement.txt
cassandra-driver
kafka-python-ng
pyspark
spark